Turn IterativeSolvers and KrylovKit into extensions
Since they are not recommended in the docs nor are they used in the default algorithm, we don't need to take the dependency.
