+++ src/iterative_wrappers.jl
@@ -218,161 +218,3 @@ function SciMLBase.solve!(cache::LinearCache, alg::KrylovJL; kwargs...)
     return SciMLBase.build_linear_solution(alg, cache.u, resid, cache;
                                            iters = cache.cacheval.stats.niter)
 end
-
-## IterativeSolvers.jl
-
-struct IterativeSolversJL{F, I, A, K} <: AbstractKrylovSubspaceMethod
-    generate_iterator::F
-    gmres_restart::I
-    args::A
-    kwargs::K
-end
-
-function IterativeSolversJL(args...;
-                            generate_iterator = IterativeSolvers.gmres_iterable!,
-                            gmres_restart = 0, kwargs...)
-    return IterativeSolversJL(generate_iterator, gmres_restart,
-                              args, kwargs)
-end
-
-default_alias_A(::IterativeSolversJL, ::Any, ::Any) = true
-default_alias_b(::IterativeSolversJL, ::Any, ::Any) = true
-
-function IterativeSolversJL_CG(args...; kwargs...)
-    IterativeSolversJL(args...;
-                       generate_iterator = IterativeSolvers.cg_iterator!,
-                       kwargs...)
-end
-function IterativeSolversJL_GMRES(args...; kwargs...)
-    IterativeSolversJL(args...;
-                       generate_iterator = IterativeSolvers.gmres_iterable!,
-                       kwargs...)
-end
-function IterativeSolversJL_BICGSTAB(args...; kwargs...)
-    IterativeSolversJL(args...;
-                       generate_iterator = IterativeSolvers.bicgstabl_iterator!,
-                       kwargs...)
-end
-function IterativeSolversJL_MINRES(args...; kwargs...)
-    IterativeSolversJL(args...;
-                       generate_iterator = IterativeSolvers.minres_iterable!,
-                       kwargs...)
-end
-
-function init_cacheval(alg::IterativeSolversJL, A, b, u, Pl, Pr, maxiters::Int, abstol,
-                       reltol,
-                       verbose::Bool, assumptions::OperatorAssumptions)
-    restart = (alg.gmres_restart == 0) ? min(20, size(A, 1)) : alg.gmres_restart
-
-    kwargs = (abstol = abstol, reltol = reltol, maxiter = maxiters,
-              alg.kwargs...)
-
-    iterable = if alg.generate_iterator === IterativeSolvers.cg_iterator!
-        !_isidentity_struct(Pr) &&
-            @warn "$(alg.generate_iterator) doesn't support right preconditioning"
-        alg.generate_iterator(u, A, b, Pl;
-                              kwargs...)
-    elseif alg.generate_iterator === IterativeSolvers.gmres_iterable!
-        alg.generate_iterator(u, A, b; Pl = Pl, Pr = Pr, restart = restart,
-                              kwargs...)
-    elseif alg.generate_iterator === IterativeSolvers.bicgstabl_iterator!
-        !_isidentity_struct(Pr) &&
-            @warn "$(alg.generate_iterator) doesn't support right preconditioning"
-        alg.generate_iterator(u, A, b, alg.args...; Pl = Pl,
-                              abstol = abstol, reltol = reltol,
-                              max_mv_products = maxiters * 2,
-                              alg.kwargs...)
-    else # minres, qmr
-        alg.generate_iterator(u, A, b, alg.args...;
-                              abstol = abstol, reltol = reltol, maxiter = maxiters,
-                              alg.kwargs...)
-    end
-    return iterable
-end
-
-function SciMLBase.solve!(cache::LinearCache, alg::IterativeSolversJL; kwargs...)
-    if cache.isfresh || !(typeof(alg) <: IterativeSolvers.GMRESIterable)
-        solver = init_cacheval(alg, cache.A, cache.b, cache.u, cache.Pl, cache.Pr,
-                               cache.maxiters, cache.abstol, cache.reltol, cache.verbose,
-                               cache.assumptions)
-        cache.cacheval = solver
-        cache.isfresh = false
-    end
-    purge_history!(cache.cacheval, cache.u, cache.b)
-
-    cache.verbose && println("Using IterativeSolvers.$(alg.generate_iterator)")
-    i = 0
-    for iter in enumerate(cache.cacheval)
-        i += 1
-        cache.verbose && println("Iter: $(iter[1]), residual: $(iter[2])")
-        # TODO inject callbacks KSP into solve! cb!(cache.cacheval)
-    end
-    cache.verbose && println()
-
-    resid = cache.cacheval.residual
-    if resid isa IterativeSolvers.Residual
-        resid = resid.current
-    end
-
-    return SciMLBase.build_linear_solution(alg, cache.u, resid, cache; iters = i)
-end
-
-purge_history!(iter, x, b) = nothing
-function purge_history!(iter::IterativeSolvers.GMRESIterable, x, b)
-    iter.k = 1
-    iter.x = x
-    fill!(x, false)
-    iter.b = b
-
-    iter.residual.current = IterativeSolvers.init!(iter.arnoldi, iter.x, iter.b, iter.Pl,
-                                                   iter.Ax, initially_zero = true)
-    IterativeSolvers.init_residual!(iter.residual, iter.residual.current)
-    iter.β = iter.residual.current
-    nothing
-end
-
-## KrylovKit.jl
-
-struct KrylovKitJL{F, A, I, K} <: AbstractKrylovSubspaceMethod
-    KrylovAlg::F
-    gmres_restart::I
-    args::A
-    kwargs::K
-end
-
-function KrylovKitJL(args...;
-                     KrylovAlg = KrylovKit.GMRES, gmres_restart = 0,
-                     kwargs...)
-    return KrylovKitJL(KrylovAlg, gmres_restart, args, kwargs)
-end
-
-default_alias_A(::KrylovKitJL, ::Any, ::Any) = true
-default_alias_b(::KrylovKitJL, ::Any, ::Any) = true
-
-function KrylovKitJL_CG(args...; kwargs...)
-    KrylovKitJL(args...; KrylovAlg = KrylovKit.CG, kwargs..., isposdef = true)
-end
-
-function KrylovKitJL_GMRES(args...; kwargs...)
-    KrylovKitJL(args...; KrylovAlg = KrylovKit.GMRES, kwargs...)
-end
-
-function SciMLBase.solve!(cache::LinearCache, alg::KrylovKitJL; kwargs...)
-    atol = float(cache.abstol)
-    rtol = float(cache.reltol)
-    maxiter = cache.maxiters
-    verbosity = cache.verbose ? 1 : 0
-    krylovdim = (alg.gmres_restart == 0) ? min(20, size(cache.A, 1)) : alg.gmres_restart
-
-    kwargs = (atol = atol, rtol = rtol, maxiter = maxiter, verbosity = verbosity,
-              krylovdim = krylovdim, alg.kwargs...)
-
-    x, info = KrylovKit.linsolve(cache.A, cache.b, cache.u; kwargs...)
-
-    copy!(cache.u, x)
-    resid = info.normres
-    retcode = info.converged == 1 ? ReturnCode.Default : ReturnCode.ConvergenceFailure
-    iters = info.numiter
-    return SciMLBase.build_linear_solution(alg, cache.u, resid, cache; retcode = retcode,
-                                           iters = iters)
-end
